refactor(db): extract provider connection deletion into leaf module
Move deleteProviderConnection, deleteProviderConnections,
deleteProviderConnectionsByProvider, reorderProviderConnections and the
shared reorderConnections helper out of the providers.ts god-file into
src/lib/db/providers/deletion.ts. Re-export from providers.ts so the
public API is unchanged; providers.ts imports reorderConnections for its
create/update reorder call sites.

Shrinks providers.ts to 943 LOC, under its file-size freeze (1034),
satisfying the shrink-only ratchet flagged by check:file-size.

Why: the #9232 purge fix pushed providers.ts past its frozen LOC,
tripping the repo file-size ratchet in CI.
